♻ refactor: ProductToggleButton 분리
